Move devicelib req mask collection into a LLVM pass
